+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ### Features
 * Add info box if build target is Android to tell users that the optimizer doesn't support Quest avatars.
 * Add validation check that checks if the optimizer got added to an optimized copy.
+* Add validation error if the amount of base playable layers isn't 5.
 
 ### Bug fixes
 * Fix extra out parameters in the fragment shader not getting added to the function declaration. [(more)](https://github.com/d4rkc0d3r/d4rkAvatarOptimizer/issues/10)